[AS7-3474] JSR88 undeployment does not work. Added more test coverage.
