+++ testsuite/integration/basic/src/test/java/org/jboss/as/test/integration/jsr88/SampleServlet.java
@@ -19,13 +19,27 @@
  * Software Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA
  * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
  */
-package org.jboss.as.test.smoke.embedded.deployment;
+package org.jboss.as.test.integration.jsr88;
 
-import java.rmi.RemoteException;
+import javax.servlet.ServletException;
+import javax.servlet.http.HttpServlet;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+import java.io.IOException;
+import java.io.PrintWriter;
 
-import javax.ejb.EJBObject;
-
-public interface Echo extends EJBObject
+/**
+ * A sample servlet
+ *
+ * @author Thomas.Diesler@jboss.com
+ * @since 05-Apr-2005
+ */
+public class SampleServlet extends HttpServlet
 {
-   String echo(String echo) throws RemoteException;
+   protected void doGet(HttpServletRequest req, HttpServletResponse res) throws ServletException, IOException
+   {
+      PrintWriter out = res.getWriter();
+      out.print("Hello World!");
+      out.close();
+   }
 }
