+++ testsuite/integration/basic/src/test/resources/jsr88/META-INF/application.xml
