+++ testsuite/integration/basic/src/test/resources/jsr88/META-INF/ejb-jar.xml
@@ -1,0 +1,19 @@
+<?xml version="1.0"?>
+<!DOCTYPE ejb-jar
+   PUBLIC "-//Sun Microsystems, Inc.//DTD Enterprise JavaBeans 2.0//EN"
+   "http://java.sun.com/dtd/ejb-jar_2_0.dtd"
+>
+
+<ejb-jar>
+   <enterprise-beans>
+      <session>
+	<ejb-name>EchoBean</ejb-name>
+	<home>org.jboss.as.test.integration.jsr88.EchoHome</home>
+	<remote>org.jboss.as.test.integration.jsr88.Echo</remote>
+	<ejb-class>org.jboss.as.test.integration.jsr88.EchoBean</ejb-class>
+        <session-type>Stateless</session-type>
+        <transaction-type>Container</transaction-type>
+      </session>
+
+   </enterprise-beans>
+</ejb-jar>
