+++ testsuite/integration/basic/src/test/resources/jsr88/WEB-INF/jboss-web.xml
