+++ testsuite/integration/smoke/src/test/java/org/jboss/as/test/smoke/jsr88/SampleServlet.java
@@ -19,7 +19,7 @@
  * Software Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA
  * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
  */
-package org.jboss.as.test.smoke.embedded.deployment;
+package org.jboss.as.test.smoke.jsr88;
 
 import java.io.IOException;
 import java.io.PrintWriter;
