+++ testsuite/integration/smoke/src/test/resources/jsr88/META-INF/application.xml
@@ -1,0 +1,16 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE application PUBLIC
+	"-//Sun Microsystems, Inc.//DTD J2EE Application 1.3//EN"
+	"http://java.sun.com/dtd/application_1_3.dtd">
+<application>
+	<display-name></display-name>
+	<module>
+		<web>
+			<web-uri>deployment-web.war</web-uri>
+			<context-root>/custom-context</context-root>
+		</web>
+	</module>
+	<module>
+		<ejb>deployment-ejb.jar</ejb>
+	</module>
+</application>
