+++ testsuite/integration/smoke/src/test/resources/jsr88/META-INF/ejb-jar.xml
@@ -8,9 +8,9 @@
    <enterprise-beans>
       <session>
 	<ejb-name>EchoBean</ejb-name>
-	<home>org.jboss.as.test.smoke.embedded.deployment.EchoHome</home>
-	<remote>org.jboss.as.test.smoke.embedded.deployment.Echo</remote>
-	<ejb-class>org.jboss.as.test.smoke.embedded.deployment.EchoBean</ejb-class>
+	<home>org.jboss.as.test.smoke.jsr88.EchoHome</home>
+	<remote>org.jboss.as.test.smoke.jsr88.Echo</remote>
+	<ejb-class>org.jboss.as.test.smoke.jsr88.EchoBean</ejb-class>
         <session-type>Stateless</session-type>
         <transaction-type>Container</transaction-type>
       </session>
