+++ testsuite/integration/smoke/src/test/resources/jsr88/META-INF/jboss-app.xml
@@ -1,0 +1,8 @@
+<?xml version="1.0"?>
+
+<!DOCTYPE jboss-app
+    PUBLIC "-//JBoss//DTD J2EE Application 1.4//EN"
+    "http://www.jboss.org/j2ee/dtd/jboss-app_4_0.dtd">
+
+<jboss-app>
+</jboss-app>
