+++ testsuite/integration/smoke/src/test/resources/jsr88/META-INF/jboss.xml
