+++ testsuite/integration/smoke/src/test/resources/jsr88/WEB-INF/badweb.xml
@@ -1,0 +1,20 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<web-app xmlns="http://java.sun.com/xml/ns/j2ee"
+  xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+  xsi:schemaLocation="http://java.sun.com/xml/ns/j2ee http://java.sun.com/xml/ns/j2ee/web-app_2_4.xsd"
+  version="2.4">
+
+  <servlet>
+    <servlet-name>SampleServlet</servlet-name>
+    <servlet-class>org.jboss.as.test.smoke.jsr88.SampleServlet</servlet-class>
+  </servlet>
+
+  <servlet-mapping>
+    <servlet-name>SampleServlet</servlet-name>
+    <url-pattern>/myapp
+/*</url-pattern>
+  </servlet-mapping>
+
+</web-app>
+
