+++ testsuite/integration/smoke/src/test/resources/jsr88/WEB-INF/jboss-web.xml
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+  <!DOCTYPE jboss-web PUBLIC
+    "-//JBoss//DTD Web Application 2.4//EN"
+    "http://www.jboss.org/j2ee/dtd/jboss-web_4_0.dtd">
+
+<jboss-web>
+  <context-root>/custom-context</context-root>
+</jboss-web>
